+++ packages/cli/src/commands/metrics/schema-api.ts
@@ -1,10 +1,27 @@
 import type Client from '../../util/client';
-import type {
-  EventSchema,
-  MeasureSchema,
-  MetricsAggregation,
-  Schema,
-} from './schema-data';
+import type { Aggregation } from '@vercel/o11y-tools/query-engine/types';
+
+export interface DimensionSchema {
+  name: string;
+  label: string;
+}
+
+export interface MeasureSchema {
+  name: string;
+  label: string;
+  unit: string;
+  aggregations: Aggregation[];
+  defaultAggregation: Aggregation;
+}
+
+export interface EventSchema {
+  description: string;
+  queryEngineEvent?: string;
+  dimensions: DimensionSchema[];
+  measures: MeasureSchema[];
+}
+
+export type Schema = Record<string, EventSchema>;
 
 const EVENT_ALIASES: Record<string, string> = {
   edgeRequest: 'incomingRequest',
@@ -52,9 +69,59 @@ function toMeasureSchema(
     name: measure.name,
     label: measure.label,
     unit: measure.unit,
-    aggregations: measure.aggregations as MetricsAggregation[],
-    defaultAggregation: measure.defaultAggregation as MetricsAggregation,
+    aggregations: measure.aggregations as Aggregation[],
+    defaultAggregation: measure.defaultAggregation as Aggregation,
   };
+}
+
+export function getEventNames(schema: Schema): string[] {
+  return Object.keys(schema).sort();
+}
+
+export function getEvent(
+  schema: Schema,
+  name: string
+): EventSchema | undefined {
+  return schema[name];
+}
+
+export function getQueryEngineEventName(schema: Schema, name: string): string {
+  const event = getEvent(schema, name);
+  return event?.queryEngineEvent ?? name;
+}
+
+export function getDimensions(
+  schema: Schema,
+  eventName: string
+): DimensionSchema[] {
+  return getEvent(schema, eventName)?.dimensions ?? [];
+}
+
+export function getMeasures(
+  schema: Schema,
+  eventName: string
+): MeasureSchema[] {
+  return getEvent(schema, eventName)?.measures ?? [];
+}
+
+export function getDefaultAggregation(
+  schema: Schema,
+  eventName: string,
+  measureName: string
+): Aggregation {
+  const event = getEvent(schema, eventName);
+  const measure = event?.measures.find(m => m.name === measureName);
+  return measure?.defaultAggregation ?? 'sum';
+}
+
+export function getAggregations(
+  schema: Schema,
+  eventName: string,
+  measureName: string
+): readonly Aggregation[] {
+  const event = getEvent(schema, eventName);
+  const measure = event?.measures.find(m => m.name === measureName);
+  return measure?.aggregations ?? [];
 }
 
 export async function fetchSchema(
